requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.0.0.post2
pytest==7.4.2
pytest-cov==4.1.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
SQLAlchemy==2.0.21
Mako==1.2.4
alembic==1.12.1
numpy==1.24.4
six==1.16.0
